# Greatly speed up doctests by compiling compatible doctests in one file