… library/test/src/types.rs
… src/doc/rustdoc/src/write-documentation/documentation-tests.md
… src/librustdoc/doctest/make.rs
… src/librustdoc/doctest/markdown.rs
… src/librustdoc/doctest/runner.rs
… src/librustdoc/doctest/rust.rs
… src/librustdoc/doctest/tests.rs
… src/librustdoc/html/markdown.rs
… src/librustdoc/passes/check_doc_test_visibility.rs
… tests/run-make/doctests-keep-binaries-2024/rmake.rs
… tests/run-make/doctests-keep-binaries-2024/t.rs
… tests/run-make/doctests-merge/doctest-2021.stdout
… tests/run-make/doctests-merge/doctest-2024.stdout
… tests/run-make/doctests-merge/doctest-standalone.rs
… tests/run-make/doctests-merge/doctest-standalone.stdout
… tests/run-make/doctests-merge/doctest.rs
… tests/run-make/doctests-merge/rmake.rs
… tests/rustdoc-ui/2024-doctests-checks.rs
… tests/rustdoc-ui/2024-doctests-checks.stdout
… tests/rustdoc-ui/2024-doctests-crate-attribute.rs
… tests/rustdoc-ui/2024-doctests-crate-attribute.stdout
… tests/rustdoc-ui/doctest/failed-doctest-should-panic-2021.rs
… tests/rustdoc-ui/doctest/failed-doctest-should-panic-2021.stdout
… tests/rustdoc-ui/doctest/failed-doctest-should-panic.stdout
… tests/rustdoc-ui/doctest/merged-ignore-no_run.rs
… tests/rustdoc-ui/doctest/merged-ignore-no_run.stdout
… tests/rustdoc-ui/doctest/wrong-ast-2024.rs
… tests/rustdoc-ui/doctest/wrong-ast-2024.stdout
… tests/rustdoc-ui/doctest/wrong-ast.rs
… tests/rustdoc-ui/doctest/wrong-ast.stdout