@Kobzol 2024-06-11
It would also be very non-trivial, implementation wise, I imagine. But what about the second approach? We could compile a single binary that looks something like this:
```rust
fn test_0() {}
fn test_1() {}

fn main() {
   let arg = args()[0];
   if arg == "test_0" { test_0(); }
   if arg == "test_1" { test_1(); }
}
```
And the rustdoc driver would then execute this binary N times (in parallel for all threads on the given machine), and run each test in a separate process.