@camelid 2024-06-11
> This doesn't seem to handle the fact that different doctests may use different `#![feature)]`s.

I think this would unfortunately break the second approach too @Kobzol