@camelid 2024-06-11
Agreed. So sounds like the idea would be:

- merge doc tests into groups based on each one's combination of edition, no_std/no_core, enabled feature gates, etc.
- create unified test for each merged group where only one test is run per execution, dependent on the args provided by rustdoc to the binary
- run each instance (i.e. provided args) of each merged test in parallel

This should obviate the need for doing this over an edition boundary since it shouldn't break any code. It would also automatically work with code that configures global state, without needing to know a `standalone` attribute. We'd also skip introducing this attribute to rustdoc since it wouldn't be necessary.