@Scripter17 2024-06-11
> However IIRC one of the slowest parts is linking
Would nightly's new linker (LLD, I think) be a relevant thing to look through for insight?