@camelid 2024-06-12
I'm not sure. It'd be worth testing. We should also look into turning off debuginfo for doctests since it's unused.