@GuillaumeGomez 2024-06-12
True, forking on linux should be pretty fast nowadays. ^^'

> But maybe it's only because the rustc test machinery has a lot of costs when starting. I wonder what would happen if we just literally did [this](https://github.com/rust-lang/rust/pull/126245#issuecomment-2161541604), or, as an alternative, set up everything with libtest, and then fork the process just before executing each test.

Hum... I'm definitely not convinced. Especially since you mentioned Windows and MacOS being slow to fork. I can provide a big source code for what rustdoc generates for `core`  later once I'm done with what I'm doing and you can experiment with forking?