@Kobzol 2024-06-12
Ok, if you can get me that large file, I'll try to play with it. I'll also try to do some experiments in the mean time.