@Kobzol 2024-06-12
Ok, turns out that I misunderstood how `rustdoc` runs the tests. It does in fact overlap building and executing these tests, and compilation is indeed the higher cost.

- `sysinfo`
   ```
   wall-time duration: 4.59s
   total compile time: 27.067s
   total runtime: 3 969s
   ```
- `core`
   ```
   wall-time duration: 102s
   total compile time: 775.204s
   total runtime: 15.487s
   ```
The total times are across all 16 threads, in my case. Looks like running the tests is just 2-10% of the total duration, at least for these two crates.

On the other hand, this makes the idea of running tests in separate processes kind of workable, because it seems like it shouldn't contribute to the total overhead that much. We'd need to benchmark it though, ofc.