@Kobzol 2024-06-12
> We should also look into turning off debuginfo for doctests since it's unused.

I checked this, and (sadly) we're already doing what we can. The tests are compiled iwthout optimization, without incremental, without debuginfo, and with `-Ccodegen-units=1`. I don't think there's much else that we can do with just flags to make this faster.