@GuillaumeGomez 2024-06-12
So most issues are solved on my side, except one:

```
memory allocation of 530632 bytes failed
```

`libtest` wants allocated data instead of working on a slice, which is problematic. Wondering if I should split doctests and call `libtest` differently or make it so that `libtest` doesn't need a `Vec` which will be modified quite a lot...