@Kobzol 2024-06-12
> Does it apply thin-local LTO? Turning that off helps with stage 0 builds, maybe it'd help with doctests too.

Isn't thin-local LTO a no-op with CGU=1? https://doc.rust-lang.org/rustc/codegen-options/index.html#lto